refactor: Update Navbar component to remove Settings link from Menubar
The Navbar component in Navbar.tsx was modified to remove the Settings link from the Menubar. This change was made to align with the updated design and functionality requirements. The Settings link was removed from the MenubarContent section of the Navbar component. This modification improves the user experience by removing unnecessary navigation options.