refactor: Update UpdateFlowValidator to make forward object optional
The UpdateFlowValidator function in chatbot.validator.ts was modified to make the forward object optional. This change allows the forward object to be omitted when validating the request body. By default, the number and message properties of the forward object are set to empty strings. This modification provides flexibility in handling the forward object in the UpdateFlowValidator function.